vix_slope_system/requirements.txt
pandas>=2.2
numpy>=1.26
yfinance>=0.2
scikit-learn>=1.4
lightgbm>=4.3
joblib>=1.4
pyyaml>=6.0
polygon-api-client>=1.12         # ← NEW for intraday minute data
